requirements.txt
flask
torch==2.2.0
timm
pillow
requests
